requirements.txt
openai 
langchain
langchain-openai
python-dotenv
qdrant-client
llama-index
pandas
html2text
trulens_eval
langchainhub